Items: Introduce InheritableItem
This is a new abstract item type that has all methods needed to query for
parents and children.
