+++ cobbler/items/item.py
@@ -17,6 +17,7 @@ V3.4.0 (unreleased):
     * Removed:
         * mgmt_classes
         * mgmt_parameters
+        * last_cached_mtime
 V3.3.4 (unreleased):
     * No changes
 V3.3.3:
@@ -101,29 +102,22 @@ V2.8.5:
 # SPDX-FileCopyrightText: Michael DeHaan <michael.dehaan AT gmail>
 
 import pprint
-import uuid
 from abc import ABC
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, Type, TypeVar, Union
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Type, TypeVar, Union
 
 from cobbler import enums, utils
-from cobbler.cexceptions import CX
 from cobbler.decorator import InheritableDictProperty, InheritableProperty, LazyProperty
-from cobbler.items.abstract.base_item import BaseItem
+from cobbler.items.abstract.inheritable_item import InheritableItem
 from cobbler.utils import input_converters
 
 if TYPE_CHECKING:
     from cobbler.api import CobblerAPI
-    from cobbler.items.distro import Distro
-    from cobbler.items.menu import Menu
-    from cobbler.items.profile import Profile
-    from cobbler.items.system import System
-    from cobbler.settings import Settings
 
 
 T = TypeVar("T")
 
 
-class Item(BaseItem, ABC):
+class Item(InheritableItem, ABC):
     """
     An Item is a serializable thing that can appear in a Collection
     """
@@ -132,113 +126,28 @@ class Item(BaseItem, ABC):
     TYPE_NAME = "generic"
     COLLECTION_TYPE = "generic"
 
-    # Item types dependencies.
-    # Used to determine descendants and cache invalidation.
-    # Format: {"Item Type": [("Dependent Item Type", "Dependent Type attribute"), ..], [..]}
-    TYPE_DEPENDENCIES: Dict[str, List[Tuple[str, str]]] = {
-        "repo": [
-            ("profile", "repos"),
-        ],
-        "distro": [
-            ("profile", "distro"),
-        ],
-        "menu": [
-            ("menu", "parent"),
-            ("image", "menu"),
-            ("profile", "menu"),
-        ],
-        "profile": [
-            ("profile", "parent"),
-            ("system", "profile"),
-        ],
-        "image": [
-            ("system", "image"),
-        ],
-        "system": [],
-    }
-
-    # Defines a logical hierarchy of Item Types.
-    # Format: {"Item Type": [("Previous level Type", "Attribute to go to the previous level",), ..],
-    #                       [("Next level Item Type", "Attribute to move from the next level"), ..]}
-    LOGICAL_INHERITANCE: Dict[
-        str, Tuple[List[Tuple[str, str]], List[Tuple[str, str]]]
-    ] = {
-        "distro": (
-            [],
-            [
-                ("profile", "distro"),
-            ],
-        ),
-        "profile": (
-            [
-                ("distro", "distro"),
-            ],
-            [
-                ("system", "profile"),
-            ],
-        ),
-        "image": (
-            [],
-            [
-                ("system", "image"),
-            ],
-        ),
-        "system": ([("image", "image"), ("profile", "profile")], []),
-    }
-
     def __init__(
         self, api: "CobblerAPI", *args: Any, is_subobject: bool = False, **kwargs: Any
     ):
         """
-        Constructor.  Requires a back reference to the CobblerAPI object.
-
-        NOTE: is_subobject is used for objects that allow inheritance in their trees. This inheritance refers to
-        conceptual inheritance, not Python inheritance. Objects created with is_subobject need to call their
-        setter for parent immediately after creation and pass in a value of an object of the same type. Currently this
-        is only supported for profiles. Subobjects blend their data with their parent objects and only require a valid
-        parent name and a name for themselves, so other required options can be gathered from items further up the
-        Cobbler tree.
-
-                           distro
-                               profile
-                                    profile  <-- created with is_subobject=True
-                                         system   <-- created as normal
-                           image
-                               system
-                           menu
-                               menu
-
-        For consistency, there is some code supporting this in all object types, though it is only usable
-        (and only should be used) for profiles at this time.  Objects that are children of
-        objects of the same type (i.e. subprofiles) need to pass this in as True.  Otherwise, just
-        use False for is_subobject and the parent object will (therefore) have a different type.
-
-        The keyword arguments are used to seed the object. This is the preferred way over ``from_dict`` starting with
-        Cobbler version 3.4.0.
+        Constructor. This is a legacy class that will be phased out with the 3.4.0 release.
 
         :param api: The Cobbler API object which is used for resolving information.
         :param is_subobject: See above extensive description.
         """
         super().__init__(api, *args, **kwargs)
 
-        self._parent = ""
-        self._depth = 0
-        self._children: List[str] = []
         self._kernel_options: Union[Dict[Any, Any], str] = {}
         self._kernel_options_post: Union[Dict[Any, Any], str] = {}
         self._autoinstall_meta: Union[Dict[Any, Any], str] = {}
         self._fetchable_files: Union[Dict[Any, Any], str] = {}
         self._boot_files: Union[Dict[Any, Any], str] = {}
         self._template_files: Dict[str, Any] = {}
-        self._last_cached_mtime = 0
-        self._is_subobject = is_subobject
         self._inmemory = True
 
         if len(kwargs) > 0:
             kwargs.update({"is_subobject": is_subobject})
-            Item.from_dict(self, kwargs)
-        if self._uid == "":
-            self._uid = uuid.uuid4().hex
+            self.from_dict(kwargs)
 
         if not self._has_initialized:
             self._has_initialized = True
@@ -553,188 +462,6 @@ class Item(BaseItem, ABC):
         except TypeError as error:
             raise TypeError("invalid fetchable files specified") from error
 
-    @LazyProperty
-    def depth(self) -> int:
-        """
-        This represents the logical depth of an object in the category of the same items. Important for the order of
-        loading items from the disk and other related features where the alphabetical order is incorrect for sorting.
-
-        :getter: The logical depth of the object.
-        :setter: The new int for the logical object-depth.
-        """
-        return self._depth
-
-    @depth.setter
-    def depth(self, depth: int) -> None:
-        """
-        Setter for depth.
-
-        :param depth: The new value for depth.
-        """
-        if not isinstance(depth, int):  # type: ignore
-            raise TypeError("depth needs to be of type int")
-        self._depth = depth
-
-    @LazyProperty
-    def parent(self) -> Optional[Union["System", "Profile", "Distro", "Menu"]]:
-        """
-        This property contains the name of the parent of an object. In case there is not parent this return
-        None.
-
-        :getter: Returns the parent object or None if it can't be resolved via the Cobbler API.
-        :setter: The name of the new logical parent.
-        """
-        if self._parent == "":
-            return None
-        return self.api.get_items(self.COLLECTION_TYPE).get(self._parent)  # type: ignore
-
-    @parent.setter
-    def parent(self, parent: str) -> None:
-        """
-        Set the parent object for this object.
-
-        :param parent: The new parent object. This needs to be a descendant in the logical inheritance chain.
-        """
-        if not isinstance(parent, str):  # type: ignore
-            raise TypeError('Property "parent" must be of type str!')
-        if not parent:
-            self._parent = ""
-            return
-        if parent == self.name:
-            # check must be done in two places as setting parent could be called before/after setting name...
-            raise CX("self parentage is weird")
-        found = self.api.get_items(self.COLLECTION_TYPE).get(parent)
-        if found is None:
-            raise CX(f'profile "{parent}" not found, inheritance not possible')
-        self._parent = parent
-        self.depth = found.depth + 1  # type: ignore
-
-    @LazyProperty
-    def get_parent(self) -> str:
-        """
-        This method returns the name of the parent for the object. In case there is not parent this return
-        empty string.
-        """
-        return self._parent
-
-    def get_conceptual_parent(self) -> Optional["BaseItem"]:
-        """
-        The parent may just be a superclass for something like a subprofile. Get the first parent of a different type.
-
-        :return: The first item which is conceptually not from the same type.
-        """
-        if self is None:  # type: ignore
-            return None
-
-        curr_obj = self
-        next_obj = curr_obj.parent
-        while next_obj is not None:
-            curr_obj = next_obj
-            next_obj = next_obj.parent
-
-        if curr_obj.TYPE_NAME in curr_obj.LOGICAL_INHERITANCE:
-            for prev_level in curr_obj.LOGICAL_INHERITANCE[curr_obj.TYPE_NAME][0]:
-                prev_level_type = prev_level[0]
-                prev_level_name = getattr(curr_obj, "_" + prev_level[1])
-                if prev_level_name is not None and prev_level_name != "":
-                    prev_level_item = self.api.find_items(
-                        prev_level_type, name=prev_level_name, return_list=False
-                    )
-                    if prev_level_item is not None and not isinstance(
-                        prev_level_item, list
-                    ):
-                        return prev_level_item
-        return None
-
-    @property
-    def logical_parent(self) -> Any:
-        """
-        This property contains the name of the logical parent of an object. In case there is not parent this return
-        None.
-
-        :getter: Returns the parent object or None if it can't be resolved via the Cobbler API.
-        :setter: The name of the new logical parent.
-        """
-        parent = self.parent
-        if parent is None:
-            return self.get_conceptual_parent()
-        return parent
-
-    @property
-    def children(self) -> List["Item"]:
-        """
-        The list of logical children of any depth.
-        :getter: An empty list in case of items which don't have logical children.
-        :setter: Replace the list of children completely with the new provided one.
-        """
-        results: List["Item"] = []
-        list_items = self.api.get_items(self.COLLECTION_TYPE)
-        for obj in list_items:
-            # TODO: Refactor once InheritableItem is available
-            if obj.get_parent == self._name:  # type: ignore
-                results.append(obj)  # type: ignore
-        return results
-
-    def tree_walk(self) -> List["Item"]:
-        """
-        Get all children related by parent/child relationship.
-        :return: The list of children objects.
-        """
-        results: List[Any] = []
-        for child in self.children:
-            results.append(child)
-            results.extend(child.tree_walk())
-
-        return results
-
-    @property
-    def descendants(self) -> List["Item"]:
-        """
-        Get objects that depend on this object, i.e. those that would be affected by a cascading delete, etc.
-
-        .. note:: This is a read only property.
-
-        :getter: This is a list of all descendants. May be empty if none exist.
-        """
-        childs = self.tree_walk()
-        results = set(childs)
-        childs.append(self)  # type: ignore
-        for child in childs:
-            for item_type in Item.TYPE_DEPENDENCIES[child.COLLECTION_TYPE]:
-                dep_type_items = self.api.find_items(
-                    item_type[0], {item_type[1]: child.name}, return_list=True
-                )
-                if dep_type_items is None or not isinstance(dep_type_items, list):
-                    raise ValueError("Expected list to be returned by find_items")
-                results.update(dep_type_items)
-                for dep_item in dep_type_items:
-                    results.update(dep_item.descendants)
-        return list(results)
-
-    @LazyProperty
-    def is_subobject(self) -> bool:
-        """
-        Weather the object is a subobject of another object or not.
-
-        :getter: True in case the object is a subobject, False otherwise.
-        :setter: Sets the value. If this is not a bool, this will raise a ``TypeError``.
-        """
-        return self._is_subobject
-
-    @is_subobject.setter
-    def is_subobject(self, value: bool) -> None:
-        """
-        Setter for the property ``is_subobject``.
-
-        :param value: The boolean value whether this is a subobject or not.
-        :raises TypeError: In case the value was not of type bool.
-        """
-        if not isinstance(value, bool):  # type: ignore
-            raise TypeError(
-                "Field is_subobject of object item needs to be of type bool!"
-            )
-        self._is_subobject = value
-
     def dump_vars(
         self, formatted_output: bool = True, remove_dicts: bool = False
     ) -> Union[Dict[str, Any], str]:
@@ -768,8 +495,8 @@ class Item(BaseItem, ABC):
         if item_dict["inmemory"]:
             for ancestor_item_type, ancestor_deps in Item.TYPE_DEPENDENCIES.items():
                 for ancestor_dep in ancestor_deps:
-                    if self.TYPE_NAME == ancestor_dep[0]:
-                        attr_name = ancestor_dep[1]
+                    if self.TYPE_NAME == ancestor_dep.dependant_item_type:
+                        attr_name = ancestor_dep.dependant_type_attribute
                         if attr_name not in item_dict:
                             continue
                         attr_val = item_dict[attr_name]
@@ -780,26 +507,6 @@ class Item(BaseItem, ABC):
                             for ancestor_name in attr_val:
                                 deserialize_ancestor(ancestor_item_type, ancestor_name)
         self.from_dict(item_dict)
-
-    def grab_tree(self) -> List[Union["BaseItem", "Settings"]]:
-        """
-        Climb the tree and get every node.
-
-        :return: The list of items with all parents from that object upwards the tree. Contains at least the item
-                 itself and the settings of Cobbler.
-        """
-        results: List[Union["BaseItem", "Settings"]] = [self]
-        parent = self.logical_parent
-        while parent is not None:
-            results.append(parent)
-            parent = parent.logical_parent
-            # FIXME: Now get the object and check its existence
-        results.append(self.api.settings())
-        self.logger.debug(
-            "grab_tree found %s children (including settings) of this object",
-            len(results),
-        )
-        return results
 
     def _clean_dict_cache(self, name: Optional[str]):
         """
